requirements.txt
Flask
Pillow
st7789 # For SPI LCD Display (requires RPi.GPIO, spidev)
spidev # Explicitly add spidev for SPI communication
numpy  # Often required by display/image libraries 